add remove/add favorite + refresh

+++ src/main/java/com/streaming/fb_streamingplatform/service/StreamingService.java
@@ -57,7 +57,8 @@ public class StreamingService {
 
     // FAVORITE
 
-    public void addFavorite(int userId, int movieId) throws Exception {
+    public List<Movie> addFavorite(int userId, int movieId) throws Exception {
+
         // Check if user exists
         if (userRepository.getByUserId(userId).isEmpty()) {
             throw new Exception("User with that ID does not exist");
@@ -68,15 +69,21 @@ public class StreamingService {
             throw new Exception("Movie with that ID does not exist");
         }
 
-        // Check if user already have that favorite movie
+        // Gets user's favorites
         List<Movie> favMovies = favoriteRepository.getFavoritesByUserId(userId);
 
+        // check if selected movie is already a favorite
         for (Movie mov : favMovies) {
             if (mov.getId() == movieId) {
-                throw new Exception("Movie is already favorite");
+                throw new Exception("Movie is already a favorite");
             }
         }
+
+        // Add favorite
         favoriteRepository.add(userId, movieId);
+
+        //Return updated favorites list
+        return favoriteRepository.getFavoritesByUserId(userId);
     }
 
     public List<Movie> getMoviesSortedByRating() {
@@ -85,31 +92,37 @@ public class StreamingService {
         } catch (SQLException e) {
             throw new RuntimeException("Failed to fetch movies sorted by rating", e);
         }
+    }
       
-        // TODO check if list is empty
-    public void removeFavorite(int userId, int movieId) throws Exception {
+
+    public List<Movie> removeFavorite(int userId, int movieId) throws Exception {
         // check if user exists
         if (userRepository.getByUserId(userId).isEmpty()) {
             throw new Exception("User with that ID does not exist");
         }
 
-        // Check if movie exists
+        // check if movie exists
         if (movieRepository.getByMovieId(movieId).isEmpty()) {
             throw new Exception("Movie with that ID does not exist");
         }
 
-        // Check if user already have that favorite movie
+        // check if user already have that favorite movie
         List<Movie> favMovies = favoriteRepository.getFavoritesByUserId(userId);
 
+        // check if user has no favorites
+        if (favMovies.isEmpty()){
+            throw new Exception("The user has no favorites to remove.");
+        }
+
+        // check if selected movie is a favorite
         for (Movie mov : favMovies) {
             if (mov.getId() == movieId) {
                 favoriteRepository.remove(userId, movieId);
-                return;
+                return favoriteRepository.getFavoritesByUserId(userId);
             }
         }
 
         throw new Exception("The user does not have the selected movie as a favorite.");
-
     }
 
     public FavoritesResult findFavoritesByEmail(String email){
